+++ packages/react-native-screen-transitions/src/shared/index.ts
@@ -71,6 +71,7 @@ export type {
 	BoundsNavigationZoomStyle,
 	GestureHandoffValues,
 	InactiveBehavior,
+	OverlayComponent,
 	OverlayProps,
 	RawGestureValues,
 	ScreenBackdropComponent,